Update README.md
I'd like to add an article I wrote on energy with a perspective that is missing from the rest of the literature. https://astigmatic.substack.com/p/bitcoin-is-the-energy-minimal-system

In addition to it being free / openly available now, I will never be paywalling this newsletter (I'll probably never paywall anything as I'm largely averse to the concept).